requirements.txt
appnope==0.1.0
decorator==4.0.9
Django==1.9.2
django-instapush==1.0.3
gnureadline==6.3.3
ipython==4.1.1
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5.1
requests==2.9.1
requests-toolbelt==0.6.0
simplegeneric==0.8.1
traitlets==4.1.0
